refactor: deal with onClickTarget in GenericTextRenderer
